@@ -8,41 +8,38 @@ authors = ["Maksim Burmistrov <burmistrovm@live.ru>"]
 start = "gunicorn -c gunicorn.py"
 
 [tool.poetry.dependencies]
-python = "^3.9"
-alembic = "^1.8"
+python = "^3.11"
+alembic = "^1.9"
+argon2-cffi = "^21.3"
 beautifulsoup4 = "^4.11"
-celery = "^5.2"
-exif = "^1.3"
-flask = "^2.1"
-Flask_SQLAlchemy = "^2.5"
-jsonschema = "^4.7"
-marshmallow = "^3.17"
-Pillow = "^9.2"
+cryptography = "^39.0"
+dramatiq = {extras = ["redis", "watch"], version = "^1.13.0"}
+exif = "^1.5"
+flask = "^2.2"
+flask-cors = "^3.0"
+flask-sqlalchemy = "^3.0"
+marshmallow = "^3.19"
+pillow = "^9.4"
 psycopg2-binary = "^2.9"
 python-dateutil = "^2.8"
-python-multipart = "^0.0"
+pyvips = "^2.2"
+pyyaml = "^6.0"
 pyzbar = "^0.1"
 requests = "^2.28"
 sqlalchemy = "^1.4"
-uvicorn = "^0.18"
-webargs = "^8.2"
-xlsxwriter = "^3.0"
-argon2-cffi = "^21.3.0"
-pyvips = "^2.2.1"
-cryptography = "^38.0.4"
-Flask-Cors = "^3.0.10"
-PyYAML = "^6.0"
+uvicorn = "^0.20"
 vk = "^3.0"
+webargs = "^8.2"
 
 [tool.poetry.dev-dependencies]
-isort = "^5.10.1"
-mypy = "^0.971"
-pylint = "^2.14.5"
-sqlalchemy-stubs = "^0.4"
-black = "^22.6.0"
-types-PyYAML = "^6.0.11"
-types-requests = "^2.28.8"
-types-python-dateutil = "^2.8.19"
+black = "^22.12"
+isort = "^5.11"
+mypy = ">=0.900,<1.0"
+pylint = "^2.15"
+types-python-dateutil = "^2.8"
+types-pyyaml = "^6.0"
+types-requests = "^2.28"
+types-sqlalchemy = "^1.4"
 
 [build-system]
 requires = ["poetry-core>=1.0.0"]
