@@ -1,14 +1,17 @@
 from collections import defaultdict
 from datetime import datetime
 from datetime import timedelta
+from datetime import timezone
 
-from celery import group
+import dramatiq
 
+from auctions.config import Config
 from auctions.db.models.auction_sets import AuctionSet
 from auctions.db.models.auctions import Auction
 from auctions.db.models.bidders import Bidder
 from auctions.db.models.bids import Bid
 from auctions.db.models.enum import CreateBidFailReason
+from auctions.db.models.enum import SortOrder
 from auctions.db.repositories.auction_sets import AuctionSetsRepository
 from auctions.db.repositories.auction_targets import AuctionTargetsRepository
 from auctions.db.repositories.auctions import AuctionsRepository
@@ -21,7 +24,6 @@ from auctions.exceptions import AuctionSetEndFailed
 from auctions.exceptions import AuctionSetStartFailed
 from auctions.exceptions import CreateBidFailed
 from auctions.services.crud_service import CRUDServiceProvider
-from auctions.services.vk_auctions_service import VKAuctionsService
 from auctions.services.vk_notification_service import VKNotificationService
 from auctions.utils import text_constants
 
@@ -30,7 +32,6 @@ class AuctionsService:
     def __init__(
         self,
         crud_service: CRUDServiceProvider,
-        vk_auctions_service: VKAuctionsService,
         vk_notification_service: VKNotificationService,
         auction_sets_repository: AuctionSetsRepository,
         auction_targets_repository: AuctionTargetsRepository,
@@ -39,9 +40,9 @@ class AuctionsService:
         bids_repository: BidsRepository,
         external_entities_repository: ExternalEntitiesRepository,
         items_repository: ItemsRepository,
+        config: Config,
     ) -> None:
         self.crud_service = crud_service
-        self.vk_auctions_service = vk_auctions_service
         self.vk_notification_service = vk_notification_service
         self.auction_sets_repository = auction_sets_repository
         self.auction_targets_repository = auction_targets_repository
@@ -50,11 +51,49 @@ class AuctionsService:
         self.bids_repository = bids_repository
         self.external_entities_repository = external_entities_repository
         self.items_repository = items_repository
+        self.config = config
 
-    def create_auction_set(self, target_id: int, amounts: dict[int, dict[int, int]]) -> AuctionSet:
+    def list_archived_auction_sets(self, target_id: int) -> list[AuctionSet]:
+        return self.auction_sets_repository.get_many(
+            AuctionSet.ended_at.is_not(None)
+            & (AuctionSet.target_id == target_id),
+            sort_key=AuctionSet.ended_at,
+            sort_order=SortOrder.DESC,
+        )
+
+    def get_active_auction_set(self, target_id: int) -> AuctionSet:
+        return self.auction_sets_repository.get_one(
+            AuctionSet.ended_at.is_(None)
+            & (AuctionSet.target_id == target_id)
+        )
+
+    def create_auction_set(
+        self,
+        target_id: int,
+        date_due: datetime,
+        anti_sniper: int,
+        item_ids: list[int],
+    ) -> AuctionSet:
         target = self.auction_targets_repository.get_one_by_id(target_id)
-        items = self.items_repository.get_random_set(amounts)
-        auction_set = self.auction_sets_repository.create(target=target, items=items)
+        items = self.items_repository.get_many(ids=item_ids)
+
+        date_due = date_due.astimezone(timezone.utc)
+
+        auction_set = self.auction_sets_repository.create(
+            target=target,
+            date_due=date_due,
+            anti_sniper=anti_sniper,
+        )
+
+        for item in items:
+            auction_set.auctions.append(
+                self.auctions_repository.create(
+                    set=auction_set,
+                    item=item,
+                    date_due=date_due,
+                )
+            )
+
         return auction_set
 
     def start_auction_set(self, set_id: int) -> AuctionSet:
@@ -63,32 +102,41 @@ class AuctionsService:
         if auction_set.started_at or auction_set.ended_at:
             raise AuctionSetStartFailed("Cannot start already started or ended auction set")
 
-        started_at = datetime.utcnow()
+        started_at = datetime.now(timezone.utc)
         auction_set.started_at = started_at
 
         for auction in auction_set.auctions:
             auction.started_at = started_at
 
-        self.vk_auctions_service.upload_auctions(auction_set)
         self.schedule_auction_set_close(auction_set)
         return auction_set
 
-    @staticmethod
-    def schedule_auction_set_close(auction_set: AuctionSet) -> None:
+    def schedule_auction_set_close(self, auction_set: AuctionSet) -> None:
         from auctions.tasks import close_auction  # pylint: disable=cyclic-import
         from auctions.tasks import close_auction_set  # pylint: disable=cyclic-import
 
-        now = datetime.utcnow()
+        now = datetime.now(timezone.utc)
 
-        (
-            group(close_auction.s(auction.id, countdown=auction.date_due - now) for auction in auction_set.auctions)
-            | close_auction_set.s(auction_set.id)
-        ).apply_async()
+        close_auction_task = dramatiq.actor(store_results=True)(close_auction)
+        close_auction_set_task = dramatiq.actor(store_results=True)(close_auction_set)
+
+        group = dramatiq.group(
+            close_auction_task.message(
+                auction.id,
+                # countdown=auction.date_due - now,
+            )
+            for auction in auction_set.auctions
+        )
+        group.add_completion_callback(close_auction_set_task.message(auction_set.id))
+        group.run()
 
     def close_auction(self, auction_id: int) -> Auction:
-        now = datetime.utcnow()
+        now = datetime.now(timezone.utc)
 
         auction = self.auctions_repository.get_one_by_id(auction_id)
+
+        if auction.started_at is None:
+            raise AuctionSetEndFailed("Cannot close not started auction")
 
         if auction.ended_at is not None:
             raise AuctionSetEndFailed("Cannot close already closed auction")
@@ -96,20 +144,27 @@ class AuctionsService:
         if auction.date_due > now:
             raise AuctionReschedule(execute_at=auction.date_due)
 
-        auction.ended_at = now
-        self.vk_auctions_service.close_auction(auction)
+        self.auctions_repository.update(auction, ended_at=now)
         return auction
 
     def close_auction_set(self, set_id: int) -> AuctionSet:
         auction_set = self.auction_sets_repository.get_one_by_id(set_id)
 
+        if auction_set.started_at is None:
+            raise AuctionSetEndFailed("Cannot close not started auction set")
+
         if auction_set.ended_at is not None:
             raise AuctionSetEndFailed("Cannot close already closed auction set")
 
-        auction_set.ended_at = datetime.utcnow()
+        auction_set.ended_at = datetime.now(timezone.utc)
 
         self.vk_notification_service.notify_winners(self.get_auction_winners(auction_set))
         return auction_set
+
+    def delete_auction_set(self, set_id: int) -> None:
+        auction_set = self.auction_sets_repository.get_one_by_id(set_id)
+        self.auctions_repository.delete(auction_set.auctions)
+        self.auction_sets_repository.delete([auction_set])
 
     @staticmethod
     def get_auction_winners(auction_set: AuctionSet) -> dict[Bidder, list[Auction]]:
@@ -123,7 +178,7 @@ class AuctionsService:
         return dict(winners)
 
     def create_bid(self, auction: Auction, bidder: Bidder, value: str) -> Bid:
-        now = datetime.utcnow()
+        now = datetime.now(timezone.utc)
 
         if auction.started_at is None or auction.ended_at is not None:
             raise CreateBidFailed(CreateBidFailReason.AUCTION_NOT_ACTIVE)
