@@ -1,5 +1,6 @@
 import re
 from datetime import datetime
+from datetime import timezone
 from string import ascii_uppercase
 from typing import Any
 from typing import Optional
@@ -397,7 +398,7 @@ class ParseService:
         if item.parse_data.get("cover_price") is None or item.parse_data.get("release_date") is None:
             return None
 
-        if item.parse_data["release_date"] + relativedelta(months=6) >= datetime.utcnow():
+        if item.parse_data["release_date"] + relativedelta(months=6) >= datetime.now(timezone.utc):
             acceptable_conditions = ["Near Mint", "Very Fine"]
             min_price_delta = 2
             to_pass = True
